airflow/requirements.txt
# Google Cloud dependencies
apache-airflow-providers-google>=10.0.0
google-cloud-bigquery>=3.0.0,<3.31.0
google-cloud-storage>=2.10.0

# Optional - helps with debugging
pandas>=1.4.0
matplotlib>=3.5.0
